% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_shifts.R
\name{all_models_predicted_change}
\alias{all_models_predicted_change}
\title{Get predicted trends from all models}
\usage{
all_models_predicted_change(some_models)
}
\arguments{
\item{some_models}{list of models produced using fit_trend_models_biomass or fit_trend_models_energy}
}
\value{
dataframe of predicted changes
}
\description{
Get predicted trends from all models in a list
}
