% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_shifts.R
\name{calc_sv}
\alias{calc_sv}
\title{Calculate state variables for each year off an ISD}
\usage{
calc_sv(isd)
}
\arguments{
\item{isd}{list with $isd, $metadata. $metadata is MATSS metadata. $isd is a dataframe with columns id (spID), mass, year, isd_seed}
}
\value{
dataframe with state variables off ISD plus location data
}
\description{
Calculate state variables for each year off an ISD
}
