% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_shifts.R
\name{compare_community_structure}
\alias{compare_community_structure}
\title{Compare sub-community properties}
\usage{
compare_community_structure(dat)
}
\arguments{
\item{dat}{a MATSS dataset}
}
\value{
df
}
\description{
Compares the mean body size, ISDs, and species composition for the first to the last 5 years of the timeseries.
}
