% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_shifts.R
\name{eval_trend_models}
\alias{eval_trend_models}
\title{Model comparison metrics}
\usage{
eval_trend_models(some_models)
}
\arguments{
\item{some_models}{result of fit_trend_models_energy or fit_trend_models_biomass}
}
\value{
dataframe of model name, aic, p values for sequential, model complexity (1-4 with 4 high)
}
\description{
Compute p values and AIC values comparing models fit with different terms. P values could be used in a kind of stepwise selection context, but I prefer to use AIC because you can compare all 4 together instead of the sequential P values.
}
