% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_shifts.R
\name{fit_trend_models_biomass}
\alias{fit_trend_models_biomass}
\title{Fit models - biomass}
\usage{
fit_trend_models_biomass(sims)
}
\arguments{
\item{sims}{result of get_annual_state_variables}
}
\value{
df
}
\description{
Fit models corresponding to syndromes of "no trend", "trend", "decoupled trend". Fits using both Gaussian and Gamma (log link) generalized linear models.
}
