% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_shifts.R
\name{get_annual_state_variables}
\alias{get_annual_state_variables}
\title{Get annual state variables}
\usage{
get_annual_state_variables(dat)
}
\arguments{
\item{dat}{MATSS dataset}
}
\value{
df
}
\description{
High level wrapper to get annual state variables for observed and null model (abundance-driven) dynamics.
}
