% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_shifts.R
\name{model_predicted_change}
\alias{model_predicted_change}
\title{Get predicted long-term change from a single model}
\usage{
model_predicted_change(one_model)
}
\arguments{
\item{one_model}{one model from the lists produced by fit_trend_models_biomass or fit_trend_models_energy}
}
\value{
data frame of predicted change, timespan, begin year, end year, correlation coefficient predicted/actual data, correlation coefficient squared, model info
}
\description{
Primarily, calculates the long-term change (ratio of last fitted value to first fitted value) from the predicted values of a model.
Using this rather than the estimate of the slope terms, because the estimated slope terms from a Gamma model are not intuitive to interpret from the readout alone.
}
