% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_shifts.R
\name{pull_focal_years}
\alias{pull_focal_years}
\title{Slice data}
\usage{
pull_focal_years(dataset, keep_years = c(1988:2018))
}
\arguments{
\item{dataset}{a MATSS-formatted dataset}

\item{keep_years}{which years to keep}
}
\value{
dataset subsetted to the years in keep_years
}
\description{
Slice data
}
