% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_shifts.R
\name{pull_model_aic}
\alias{pull_model_aic}
\title{Extract AIC for one model}
\usage{
pull_model_aic(one_model)
}
\arguments{
\item{one_model}{a GLM from the list of models returned from fit_trend_models_biomass or fit_trend_models_energy}
}
\value{
data frame with model family, link function, formula, dataset name, AIC, AICc
}
\description{
Extract AIC for one model
}
