% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_shifts.R
\name{relabund_null_model}
\alias{relabund_null_model}
\title{Null model of no size change}
\usage{
relabund_null_model(dataset, resp_seed = 1989)
}
\arguments{
\item{dataset}{a MATSS dataset}

\item{resp_seed}{a seed}
}
\value{
df
}
\description{
Resamples species' abundances according to their mean relative abundance across the entire timeseries.
}
\details{
This is the null model to characterize "no change in the size structure" by keeping species composition fixed to the long-term average across the timeseries.
}
